+++ functions/public/Get-PSWorkItem.ps1
@@ -84,11 +84,15 @@ Function Get-PSWorkItem {
         StartTimer
         $PSDefaultParameterValues['_verbose:Command'] = $MyInvocation.MyCommand
         $PSDefaultParameterValues['_verbose:block'] = 'Begin'
-        _verbose -message $strings.Starting
-        _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
-        _verbose -message ($strings.UsingModule -f (Get-Command -name $MyInvocation.MyCommand).Version)
-        _verbose -message ($strings.UsingDB -f $path)
-        _verbose ($strings.DetectedCulture -f (Get-Culture))
+        if ($MyInvocation.CommandOrigin -eq 'Runspace') {
+            #Hide this metadata when the command is called from another command
+            _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
+            _verbose -message ($strings.UsingHost -f $host.Name)
+            _verbose -message ($strings.UsingOS -f $PSVersionTable.OS)
+            _verbose -message ($strings.UsingModule -f $ModuleVersion)
+            _verbose -message ($strings.UsingDB -f $path)
+            _verbose ($strings.DetectedCulture -f (Get-Culture))
+        }
     } #begin
 
     Process {
@@ -105,7 +109,7 @@ Function Get-PSWorkItem {
                 also simplifies things when running under different cultures.
                 "days" {
                         $d = (Get-Date).AddDays($DaysDue)
-                        $query = "Select * from tasks where duedate <= '$d' collate nocase"
+                        $query = "Select * from tasks where DueDate <= '$d' collate nocase"
                     }
             #>
             'id' { $query = "Select * from tasks where ID ='$ID'" }
@@ -129,9 +133,9 @@ Function Get-PSWorkItem {
                 _verbose -message ($strings.FilterItemsDue -f $d)
                 <#
                 This format doesn't appear to respect culture
-                $tasks = ($tasks).Where({ [DateTime]$_.duedate -le $d })
+                $tasks = ($tasks).Where({ [DateTime]$_.DueDate -le $d })
                 #>
-                $tasks = $tasks | Where-Object { ($_.duedate -as [DateTime]) -le $d }
+                $tasks = $tasks | Where-Object { ($_.DueDate -as [DateTime]) -le $d }
                 _verbose -message ($strings.Refiltering -f $tasks.count)
             }
             $i = 0